drop the provider; keep the shared context type
The provider de-duplicated three props at one real call site, at the
cost of four exports, a precedence rule, and making currentUserId and
resolveAuthor optional (so forgetting them was no longer a type error).

Keep the part that answered the feedback — one CommentingContext in
place of three copies of the same interface — and let hosts spread one
object into both surfaces. A provider stays additive if a third surface
ever earns it.
