+++ packages/commenting/src/canvas/comments-sidebar.tsx
@@ -12,7 +12,7 @@ import { CommentListItemProps, CommentsList } from '../ui/comments-list'
 import { UNKNOWN_COMMENT_AUTHOR } from './comment-render'
 import { CommentsFilterMenu } from './comments-filter-menu'
 import { CommentsVisibilityToggle } from './comments-visibility-toggle'
-import { type CommentingContext, useCommentingContext } from './context'
+import { type CommentingContext } from './context'
 import { useComments, useCommentThreads } from './hooks'
 import { useCommentingEnabled } from './license'
 import { useCommentingOptions } from './options'
@@ -21,15 +21,16 @@ import { commentsSidebarOpen, openThreadId, sidebarFilters } from './state'
 import { focusThread } from './thread-state'
 
 /**
- * The host wiring for {@link CanvasCommentsSidebar}: the {@link CommentingContext} fields it
- * reads, each falling back to the enclosing {@link CommentingProvider} when left unset, plus the
- * panel's own slots. A known `currentUserId` enables the "only your threads" filter, and an
- * `isCommentUnread` the "only unread" one.
+ * The host wiring for {@link CanvasCommentsSidebar}: the {@link CommentingContext} fields it reads,
+ * plus the panel's own slots. A non-null `currentUserId` enables the "only your threads" filter, and
+ * an `isCommentUnread` the "only unread" one. `CanvasComments` takes the same fields, so a host
+ * mounting both can spread one object into each.
  *
  * @public
  */
-export interface CanvasCommentsSidebarProps extends Partial<
-	Pick<CommentingContext, 'currentUserId' | 'resolveAuthor' | 'isCommentUnread'>
+export interface CanvasCommentsSidebarProps extends Pick<
+	CommentingContext,
+	'currentUserId' | 'resolveAuthor' | 'isCommentUnread'
 > {
 	/** Header above the list. */
 	header?: ReactNode
@@ -44,12 +45,8 @@ export interface CanvasCommentsSidebarProps extends Partial<
  * or always-on list.
  * @public @react
  */
-export function CanvasCommentsSidebar({
-	header,
-	empty,
-	...contextProps
-}: CanvasCommentsSidebarProps) {
-	const { resolveAuthor, currentUserId, isCommentUnread } = useCommentingContext(contextProps)
+export function CanvasCommentsSidebar(props: CanvasCommentsSidebarProps) {
+	const { resolveAuthor, currentUserId, isCommentUnread, header, empty } = props
 	// Name-only view of the resolver, for the plaintext previews (which resolve @-mentions).
 	const resolveName = useCallback((id: string) => resolveAuthor(id)?.name, [resolveAuthor])
 	const editor = useEditor()
